Stop linking directories with unrelated host-specfic overrides
The installation script will now create all the target directories that
contains any host-specific configuration. This is done to prevent it
from linking those directories directly under home. Now it rather links
the directory content file by file, so that it will not expose
configuration files for unrelated hosts.

In e.g. .Xresources.d it have been troublesome when the scripts links
this directory directly, so that the DPI setting for a particular host
have been installed on every other host also.