Stop using ninja as the default bmake generator
There is a [ninja #1251] issue that forces cmake to generate build files
with relative instead of absolute paths. The effect is that debug
information in binaries are relative to the build directories, which for
out-of-source builds is not very compatible with running vim from the
project root. When running vim from the project root the paths reported
for build errors, test failures, and coverage information is wrong.

[ninja #1251]: https://github.com/ninja-build/ninja/issues/1251