Fix e.g. /usr/sbin overriding /usr/local/bin PATH
Prepending */sbin paths worked for the intra-sbin prioritization, but
unfortunately it now messed with their order relative to the */bin
companions. The result is that /usr/sbin was given a higher priority than
/usr/local/bin, which is unfortunate, and particularly in archlinux it
is a disaster because /usr/sbin is a symlink to /usr/bin, so it became
impossible to install any /usr/local overrides for installed binaries.

Here it is fixed by prepending both sbin and bin companions to the PATH,
and also written quite a bit more compact with the help of extended
filename expansion.