Replace ranger setlocal setting with setinpath/-regex
Because setlocal became an alias for the new setinpath setting somewhere
between ranger 1.9.3-9 and ranger-git 1.9.3.719.g136416c7-1. The old
behavior was more like setinregex when the path was specified. Use
setinregex when the point is to recurse and setinpath where it is more
useful to only work on a given directory (e.g. download).