Fix git hunk-header for [[nodiscard]] C/C++ functions
Or functions stating with any other attributes. This is based on the
function names/... part of the [default pattern], but with the added '['
character to allow for C++ attributes. I have no clue why anyone would
want to match the other parts of the default pattern, so it is nice to
just drop them.

[default pattern]: https://github.com/git/git/blob/bcb6cae2966cc407ca1afc77413b3ef11103c175/userdiff.c#L76