Use nvim K for man-pages regardless
Removing the K hotkey as suggested in the `lsp-defaults` manual crashes
for ruby, so stop doing that. The `lsp-defaults` should avoid overriding
the K hotkey (and currently does so) when a custom (buffer-local)
keywordprg is set, so make sure we always use `:Man`. However, it is
supposed to do that for non-default keywordprgs, and `:Man` is usually
the default, so who knows how long this will work.